% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dummy_fwdControl_generator}
\alias{dummy_fwdControl_generator}
\title{Dummy fwdControl object creator}
\usage{
dummy_fwdControl_generator(years = 1:10, niters = 10)
}
\arguments{
\item{years}{numeric vector of years in the control object}

\item{niters}{the number of iterations}
}
\value{
A fwdControl object
}
\description{
Creates a dummy fwdControl object for testing purposes
}

