% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{fwdControl-class}
\alias{fwdControl-class}
\title{fwdControl}
\description{
A class to hold the parameters for running a projection using \code{fwd}.
Add something about the constructors
}
\details{
Something about the columns and checks
}
\section{Slots}{

\describe{
   \item{\code{target}:}{A \code{data.frame}. See details below.}
   \item{\code{target_iters}:}{An array with the values of the targets (\code{min_value}, \code{value} and \code{max_value} by iteration.}
    }
}

